chore(gui-components): move date and time utils to internals
