+++ libs/gui/components/src/index.ts
@@ -53,38 +53,3 @@ export type { WeekInfo } from './lib/utils/week-info';
 // ─── Controllers ───
 
 export { GUIAriaController } from './lib/controllers/aria.controller';
-
-// ─── Date utilities ───
-
-export {
-  createDateRange,
-  getDateFormatParts,
-  getDayLabel,
-  getMonthName,
-  getMonthYearParts,
-  getOrderedWeekDays,
-  getWeekdayLabels,
-  isDateInVisibleMonths,
-  isSameDay,
-  isToday,
-  maxValidDayInMonth,
-  mergeDateRanges,
-  toISODateString,
-  weekDaysOrder,
-} from './lib/utils/date';
-
-// ─── Time utilities ───
-
-export {
-  from24Hour,
-  getDateTimeFormatParts,
-  getDayPeriodLabels,
-  getTimeFormatParts,
-  parseISODateTimeString,
-  parseISOTimeString,
-  resolveHourFormat,
-  to24Hour,
-  toISODateTimeString,
-  toISOTimeString,
-  type HourFormat,
-} from './lib/utils/time';
