+++ libs/gui/components/src/internals.ts
@@ -1,4 +1,33 @@
 export { createIntersectionObserver } from './lib/components/tabs';
+export {
+  createDateRange,
+  getDateFormatParts,
+  getDayLabel,
+  getMonthName,
+  getMonthYearParts,
+  getOrderedWeekDays,
+  getWeekdayLabels,
+  isDateInVisibleMonths,
+  isSameDay,
+  isToday,
+  maxValidDayInMonth,
+  mergeDateRanges,
+  toISODateString,
+  weekDaysOrder,
+} from './lib/utils/date';
 export { addErrors, addHint, addIcon, addLabel } from './lib/utils/templates';
 export type { ControlTemplateData } from './lib/utils/templates';
+export {
+  from24Hour,
+  getDateTimeFormatParts,
+  getDayPeriodLabels,
+  getTimeFormatParts,
+  parseISODateTimeString,
+  parseISOTimeString,
+  resolveHourFormat,
+  to24Hour,
+  toISODateTimeString,
+  toISOTimeString,
+  type HourFormat,
+} from './lib/utils/time';
 export type { TabsEventDetail, AccordionEventDetail } from './lib/widget-event.details';
